requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
black==21.6b0
capstone==4.0.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colored-traceback==0.3.0
cryptography==3.4.7
greenlet==1.1.0
idna==2.10
intervaltree==3.1.0
isort==5.9.1
Mako==1.1.4
MarkupSafe==2.0.1
msgpack==1.0.2
mypy-extensions==0.4.3
packaging==20.9
paramiko==2.7.2
pathlib2==2.3.5
pathspec==0.8.1
plumbum==1.7.0
psutil==5.8.0
pwntools==4.5.1
pycparser==2.20
pyelftools==0.27
Pygments==2.9.0
PyNaCl==1.4.0
pynvim==0.4.3
pyparsing==2.4.7
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.1
regex==2021.4.4
requests==2.25.1
ROPGadget==6.5
rpyc==5.0.1
six==1.16.0
sortedcontainers==2.4.0
toml==0.10.2
unicorn==1.0.2rc3
urllib3==1.26.6
